Separate protobuf generation from the C++ build.
Since the authenticator.proto file now lives in the obj-endpoint repository,
there's really no justification for tieing the protobuf generation for the
Python code to the C++ code. Copy and edit the buf generation code from the
C++ and update the README.

So now there's no copy-the-generated-code step. It's already here.

There are instructions for regenerating the Python protobuf code in the README
too, though you'll probably need to do a little work to get the tools
installed. It could easily be dockerised if people are struggling.